requirements.txt
aniso8601
asyncio
attrs
autopep8
boto3
botocore
certifi
charset-normalizer
chevron
click
dukpy
Flask
Flask-Cors
Flask-JWT-Extended
Flask-RESTful
flask-restx
Flask-SQLAlchemy
greenlet
gunicorn
idna
importlib-metadata
inflect
itsdangerous
javascripthon
Jinja2
jmespath
jsonschema
MarkupSafe
marshmallow
marshmallow-enum
packaging
Pillow
psycopg2-binary
py2ng
pycodestyle
PyJWT
PyMySQL
pyparsing
pyrsistent
python-barcode
python-dateutil
pytz
requests
s3transfer
six
spark-parser
sqlacodegen
SQLAlchemy
toml
urllib3
Werkzeug
xdis
zipp
